x86/mm: Widen the x86 paging get and set functions to unsigned long

- hap_set_allocation()/hap_get_allocation() take/return unsigned long
  - the local 'pg' in hap_get_allocation() is widened so
         return ((pg >> (20 - PAGE_SHIFT))
    no longer truncates above 16TB.

- shadow_set_allocation()/shadow_get_allocation()/sh_min_allocation()
  use unsigned long for pages.

- paging_set_allocation() takes unsigned long.

- arch_set_paging_mempool_size(): drop the pages != (unsigned int)pages
  guard that previously rejected pools above 16TB.

+++ xen/arch/x86/include/asm/hap.h
@@ -33,8 +33,8 @@ int   hap_track_dirty_vram(struct domain *d,
                            XEN_GUEST_HANDLE(void) guest_dirty_bitmap);
 
 extern const struct paging_mode *hap_paging_get_mode(struct vcpu *v);
-int hap_set_allocation(struct domain *d, unsigned int pages, bool *preempted);
-unsigned int hap_get_allocation(struct domain *d);
+int hap_set_allocation(struct domain *d, unsigned long pages, bool *preempted);
+unsigned long hap_get_allocation(struct domain *d);
 
 #endif /* XEN_HAP_H */
